fix structure of signalR and notification api
