+++ src/const/notification.ts
@@ -1,0 +1,3 @@
+export type NotificationType = "Order" | "Promotion" | "Product" | "General" | "System";
+
+export type RoleName = "Customer" | "Seller" | "Staff" | "Admin";
